+++ modules/portal-api/src/main/java/com/enonic/xp/portal/url/PortalUrlBuilders.java
@@ -28,11 +28,6 @@ public final class PortalUrlBuilders
         return builder;
     }
 
-    public GeneralUrlBuilder generalUrl()
-    {
-        return new GeneralUrlBuilder().baseUri( this.context.getBaseUri() );
-    }
-
     public AssetUrlBuilder assetUrl()
     {
         return defaults( new AssetUrlBuilder() ).module( this.context.getModule() );
@@ -64,32 +59,8 @@ public final class PortalUrlBuilders
     }
 
     @Deprecated
-    public GeneralUrlBuilder createUrl( final String path )
-    {
-        return generalUrl().contentPath( path );
-    }
-
-    @Deprecated
-    public AssetUrlBuilder createResourceUrl( final String path )
-    {
-        return assetUrl().path( path );
-    }
-
-    @Deprecated
-    public ImageUrlBuilder createImageUrl( final String name )
-    {
-        return imageUrl().imageName( name );
-    }
-
-    @Deprecated
     public ImageUrlBuilder createImageByIdUrl( final ContentId contentId )
     {
         return imageUrl().imageId( contentId );
-    }
-
-    @Deprecated
-    public ServiceUrlBuilder createServiceUrl( final String name )
-    {
-        return serviceUrl().service( name );
     }
 }