+++ modules/portal-xslt/src/main/java/com/enonic/xp/portal/xslt/impl/function/XsltFunctionLibrary.java
@@ -6,7 +6,6 @@ public final class XsltFunctionLibrary
 {
     public void registerAll( final Configuration config )
     {
-        config.registerExtensionFunction( new GeneralUrlFunction() );
         config.registerExtensionFunction( new PageUrlFunction() );
         config.registerExtensionFunction( new ImageUrlFunction() );
         config.registerExtensionFunction( new AssetUrlFunction() );